Allow specification of ordering of the base path in directory sorting algo.
This allows different commands to order their directories in different ways.

The `cd` command assumes you are not trying to change directory into the
directory you are already in, and hence sorts such that the current directory is
always at the bottom.

The `browse` command, when passed a slug, assumes that you are more likely want
the current clone, and orders the current working directory at the top.